Start CertificateRotation in IHostedService instead from IApplicationBuilder, move AddMutualTls methods to better named static class
